go('check') was failing on MacOS (#35)
With clang on MacOS, using strcpy to copy to overlapping memory segments
causes a core dump. (Abort trap: 6). The go macro was calling
setvalue4name, which called setNameValuePair(). This called
strcpy() which copied one portion of a string to another portion
of that same string. There may be other parts of VJ with similar
issues.