MacOS PSG libraries have a couple of problems
I am not sure why, but the -mmacosx-version-min option to clang
causes the ios_base4Init symbol not found error.

Replacing CXX=clang++ in the cppEnv selects clang++ rather than g++.

The libparam.dylib was not being used because it's install_name was
not correct. "otool -D /vnmr/lib/libparam.dylib" showed its name
as /vnmr/lib/libpsglib.dylib.  The cause is that all "replace" calls
on an Environment in SConstruct occur before the Environment is used.
The second replace on LINKFLAGS over-wrote the first replace on LINKFLAGS.

Again, for reasons I do not entirely understand, inclusion of x_ps.o in
the libpsglib.dylib selects that x_pulsesequence() function instead of
the one compiled with the psglib entry. With static libraries (libpsglib.a),
the one in the pulse sequence is selected. I just removed x_ps.o from
libpsglib.dylib.

For fixpsg and psggen, I added targets to make the dylibs in makeuserpg.lnx.