FIxed installation scripts (#122)
* FdfToDcm does not compile on MacOS

Fixed the SConstruct so it is only built on Linux.
It uses old-style #include <backward/strstream> rather than

* Fixed scripts used to install OpenVnmrJ

Installation was generating error messages. The standard sudo is now
used. Linux obsoleted kudzu so we no longer need to deal with it.
Also copy some rpms to the dvd for installation (tcl and tk)