Issue 141 fixed (#143)
* Fix for issue 137 3D Gradient shimming failure (#138)

The gxyzautoshim macro will fail with $isvnmrj variable undefined
The bug was introduced in VJ 4.0. Reported in SpinSights.

* Issue141 (#142)

* Fix for issue 137 3D Gradient shimming failure (#138) (#139)

The gxyzautoshim macro will fail with $isvnmrj variable undefined
The bug was introduced in VJ 4.0. Reported in SpinSights.

* Fix for issue 141. Renamed explist files to expqueue

The fidlib files for CRAFT included some obsolete explist files in
dirinfo/macdir. These are renamed to expqueue. Also updated table.c
for a couple of commands under development.