makefid could fail (#150)
If makefid is called repeatably, such as when making FIDs for a 2D
data set, it could fail. It counted lines in the text file by calling
the external "wc -l" command. This involves forking, waiting for child
to finish, catching signals, etc. Replaced with opening the file
and counting newlines. Faster and avoids the problems with popen.