fidmax and dpcon updated
fidmax could fail if makefid is used to make a new FID.
dpcon could not be cancelled. It now looks at the interuption
flag. writeSocket() needed to be fixed in case the interruption
occurred while it was writing a message to the java front end.
The java front end requires a complete message.