Fixed function signatures (#193)
The readfile command was failing on one PC but worked correctly on
two others. The failure was in the rfindVar section. It claimed it
could not find a local $ variable, but it was absolutely present.
Not clear what the problem is but I noticed the function signatures
for some of the things used by rfindVar (findName, etc) were  different.
These modifications made it work on all three systems.