Fixed a Viewport bug and a Vnmrbg crash
For the viewport bug, just need to check for a null pointer before setting
the "busy" cursor. This error showed up in a VnmrjMsgLog when changing
viewports.
To reproduce the Vnmrbg core dump, process 2D data in exp4, then
process 1D data in exp5 and display the spectrum (ds). jexp4 and then
reposition the command line. Vnmrbg will generate "Data file is not open
currently" messages and core dump.