Updates to vnmrj.jar (#279)
Added "rebuild" option to xml popups. If the text file on which the popup is based
is changed, setting the "expanded" field to rebuild and changing a parameter
it is watching (vq field) will cause the popup to re-read the text file.
Also, avoid doing unneeded work it the Locator is turned off.