update for OpenVnmrJ loss of /usr/varian, keep old log files
This fixes #294.  It was checking for existence of /usr/varian to
verify that ovj had been installed, but since that's gone, now
checks for existence of /vnmr/templates.

Also added some code to roll log files instead of over-writing,
up to 10, then just keep over-writing 10.  Good enough for now,
but true rolling would be better.