fpBrowser can write selections to a file (#314)
If too many selections are made, the call to the macro can fail
because the string is too long. Passing a fifth argument, a path name,
will cause fpBrowser to write the selections to that file. It will
then call the macro with the arguments 'path' and the path name
supplied as the fifth argument to fpBrowser.