gxyzmapsys could fail (#364)
Fix provided by Bert Heise. His comments:
The problem is that if a previous map doesn't exist (maybe it got interrupted
before it could be saved or whatever - the map in quesiton is in the list of
available maps but the file isn't there…), gxyzmapsys has nowhere to
go - gxyzrtmap aborts and mapsys has no chance of launching normally in this state.
Modified gxyzmapsys to check for the presence of the map before calling rtmap
and when not finding a map, call gmapxyz directly instead.