Substr (#417)
* Added option to substr basename

Suppling a suffix with one return argument is now consistent
with the Linux basename command. The suffix is removed from
the basename. With two return arguments, the behavior has
not changed.

* The basename test to remove a suffix was not correct.

It would fail if the suffix was also part of the basename.
For example, substr('myfid_fid','basename','fid') did not
remove the last three characters (fid). It now gives
myfid_ as it should.