Improved the OpenVnmrJ installer
When someone ran ./load.nmr and no existing OVJ / VJ
was installed, the installer selected the first tab (ProPulse)
which in most cases is wrong. Several peole have installed
ProPulse, leading to an unusable system. The installer now
comes up with a message to choose the correct system. The
title bar also gives feedback on the system that will be
installed.