cryo_noisetest was failing
cryo_noisetest was using the side effect of the purge command
to abort the macro at certain stages. With the change that purge
no longer caused an abort, cryo_noisetest failed. The calls to purge
are replaced with "return" statements. Also removed shell calls and
some fixes to cryo_cndspec