add procdim to appendlist only when it changes
every time ft is called, procdim was being added to the
list of parameters that changed, causing java to update
items that depend on it. This caused quite a bit of overhead
during mtune, which calls ft every BS. Now it is appended
only if it changes.
Added Proclog to turn on debugging for Procproc
Some compiler warning cleanup