Added 'grep -iw' option to append (#525)
Fixed a bug in append where it would fail a word search if the
word was at the tail end of a longer word. For example
appendStr('another string','grep -w','other')
would find 'other' and the end of 'another' and pass the
grep -w test. Also, if the file is a dos file and the search
word is at the end of the line, it would fail the test.
Updated the man page and fixed all compiler warnings.