Fix append bugs. New options for systemtime and mv
If a filter after the head filter removes a
line, then the number of lines passed by the head
filter is increased. So if 'head',4 were used and a
subsequent 'tail',1 removed a line, the head would
act like it was called with 'head',5
The '|wc w' filter had a typo. It actually checked
against '|wc c'.
The systemtime now supports a third argument to reformat
a previously formatted date string. Updated the man page.
The mv command can move the contents of a directory using
fast library calls rather than calling the external mv
command as in mv('directory/*','newdirectory')