OVJ was not working if java newer than 1.8 (#575)
The CORBA modules have been removed from java 9 and newer.
The LC-NMR parts of OVJ requires them. Now do a run-time
test of the java version and only load the CORBA stuff
if java is 1.8 or older. Also fixed a typo in vnmrj script.
This fix also works on MacOS.