added tools needed to fix Study Queue bug
There is a race between java building the SQ from an study.xml
file and Vnmrbg constructing another study.xml based on a change
from an automation run. This results in a SAX parse error.
Added tools that will be used to fix this long-standing bug. One tool,
ovjCount is documented. The fix is to make a copy of the study.xml
and have java read and delete the copy.