xml files could leave $variables in current tree (#701)
If the Vnmr Command field (vc) of an xml file contained a $variable,
that $variable would be created in the current parameter tree.
If another xml file used the same $variable but as a different
type (string vs real), an error would occur. OVJ now wraps the command
field in a jFunc command so that the parameter stack can be handled
correctly (push and pop).