spExec failed with a write error (#731)
Lookup was returning 'inf' as a numeric value. This caused the write
command to fail. Added a test to check the values before trying to
write them. However, format with the 'isreal' option treats 'inf' as
a real. Added 'isfinite' option to format which returns that 'inf' is
not a finite number. Documented format changes.