Fix all the off by one index errors around sequence names
There is a lot of testing in the nvpsg .h files against name
length > NSUFFIX. This is the size of the internal structures
that store the names. But the test is forgetting the loss of
one char to store the NULL char at the end of the string. So
I've made all tests fail if room is not left for the string
and its terminating NULL char. Otherwise a string of len
NSUFFIX will overwrite memory.