After studying RFController code more implement a possible fix
This is my best understanding of how the original developers
meant to implement the setUserGate() code. You can see it is only
called from two places. In one place the "which" variable is
always between 1 and 3 and so a single gate op. The other place
sets "which" to 7 and passes 0 to do a clear all. But there is no
implication that "which" == 7 implies a clear all. One could set
all the gates on if desired. This code departs from how it has
been running forever. It is my belief that the running code is
buggy. I am implementing this fix on a declined pull request in
case we change our minds later.