Fix waveform generation code to avoid controller board timing issues
Dan, we are continuing to test my changes to avoid the ticks problem.
With my fixes in place we have not been able to recreate the bug
anymore through a total of 15,000 iterations of our test program.
It used to fail 1 in 100 or 1 in 50 iterations. This is a PR with
our current fix. We can wait for further testing before merging.
We will be doing some of that testing over the next few weeks.

The fix is basically to change the "num accum pulses >= 255" to
">= 5" in all the places it occurs in solidchoppers.h (and similar
changes to solidshapegen.h). This prevents the 1st column of the
shape file from ever exceeding 450 (which I think stands for
0/90/180/270/360 biased by 90 because the 1st column is never <=
that). It seems somewhere in the (hardware side?) code there is
an assumption that the 1st column values are normalized to 0/360
and the routines in solidchoppers.h don't do that. Maybe we can
track this issue down in the code? Perhaps there is a more elegant
fix.

I found 5 to be the best value by doing a binary search on 0 to
255 and continuing until the bug went away. The values 1-5 all
work. Values 6 and greater fail.

As an aside I know the 1st column is a timing but it seems like
maybe somewhere it is getting treated as an angle.