Infostat could fail
It could core dump and/or write random values into the log files if
the -l option is used. It only fails if compiled with newer versions
of gcc (eg, gcc 8). Also, compiling as 64-bit does not show the problem,
only 32-bit fails. It seems related to the new -fno-omit-frame-pointer
option. Compiling with no optimization (-O0) or with -O2 works. Only
using -O (eg -O1) fails.