Updated ovjBuild and some installation scripts (#841)
ovjBuild requires git to function. It will be installed
if needed. Added -c (--codeonly) option to just clone the
repositories. Documented changes.
On some Ubuntu systems, it seems that the /etc/lsb-release file
may not exist. Generalized the mechanism to determine the release
version to use lsb-release, os-release, or the hostnamectl command.