A scons issue caused vnmrbg compile to fail
This happened on Ubuntu 24.04. During the vnmrbg compile, files
from the Asp directory are linked. The scons tried to compile some
of them before the link was complete. This did not happen if
"scons -j 1" is used. Also found that moving the compile of vnmrbg
to the start of the compilation list seems to fix it.