Modification to makeuser. Add fixuser script
makeuser no longer removes files from persistence, seqlib, or
psg user directories. It will update appdir_* files in persistence
if it finds an obsolete version of CPpatch. The fixuser script
is added to do those things that are removed from makeuser.
The fixuser script will generally only be used if something seems
to have gone wrong. A manual for fixuser is added.