dpcon failed on AlmaLinux 9.4
I think this is a bug in AlmaLinux 9.4. It works in AlmaLinux 9.2.
dpcon sends data to java (vnmrj.jar) over a socket. java tests
for the number of bytes with the available() method of DataInputStream.
If fewer than expected bytes are available, it waits until they
are available.  On Linux 9.4, it waits forever. The fix is to read the
number of bytes that are available and then check to see if more arrived.