Updates for using NMRPipe
Uses better bindir directory for MacOS. Added option to
ovjGetpipe to specify the binType to use. Also makes a link
in nmrpipe/com to the proper nmrInit*com file, which vnmrpipe
script uses. This link may be a temporary fix, since the official
NMRPipe mght do it in the future.