The copy command could fail.
If the to_file does not exist, the logic could decide the
from_file and to_file were the same. It depends an how the
buf variable is initialized when the stat call fails. Now also
checks the result of the stat call, which will be -1 if the file
does not exist.