atcmd active keyword re-defined (#984)
It used to mean the OVJ that initially called the atcmd. It now
means the OVJ that initially called atcmd or any currently running
foreground OVJ started by the same user. An atcmd entry called with
the 'active' keyword will not be removed from the atcmd queue if there
is no OVJ to execute it.