On AlmaLinux 9, java may not be installed
If upgraded from Alma 8.10, the test to check if java needs to be
installed may fail, due to the presence of java-1.8.0-openjdk.
A java-17-openjdk-headless became the active version of java.
It now tries to install java-17-openjdk  in all cases. If it is
already installed, yum will just skip it. Thanks to Eugenio
Alvarado for help with this problem.